requirements.txt
python-dotenv==1.0.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

streamlit==1.33.0
requests==2.31.0

langchain==0.2.11
langchain-openai==0.1.7
langchain-community==0.2.10
openai==1.44.0
chromadb==0.5.3

tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2